docs(icon): Refactoring icon story organization
