+++ packages/icon/src/__stories__/iconKeys.js
@@ -1,0 +1,197 @@
+import iconSizes from "../sizes";
+
+/** @todo Generate this from build script */
+const iconKeys = {
+  [iconSizes.PX_24]: [
+    "add",
+    "archive",
+    "assets",
+    "attachment",
+    "back",
+    "bookmark",
+    "building-ops",
+    "calendar",
+    "caret",
+    "check-white",
+    "checklist",
+    "checkmark-blue-dark",
+    "checkmark",
+    "clear-small",
+    "clock",
+    "close-hamburger",
+    "close-small",
+    "close",
+    "collaboration",
+    "collapse-panel",
+    "collapse",
+    "comment",
+    "compare",
+    "complete",
+    "copy",
+    "cost-control",
+    "document-management",
+    "double-caret",
+    "download",
+    "edit",
+    "error",
+    "expand-panel",
+    "expand",
+    "export",
+    "external",
+    "external-link",
+    "favorite",
+    "field",
+    "file-assembly",
+    "file-document",
+    "file-generic",
+    "file-image",
+    "file-part",
+    "file-pdf",
+    "file-presentation",
+    "file-spreadsheet",
+    "file-video",
+    "file-zip",
+    "filter-tokens",
+    "filter",
+    "folder-open",
+    "folder",
+    "forward",
+    "globe",
+    "grid",
+    "hamburger",
+    "help",
+    "hidden",
+    "home",
+    "info",
+    "insight",
+    "issue",
+    "item",
+    "layout",
+    "list",
+    "location",
+    "lock",
+    "model-coordination",
+    "more-vertical",
+    "paste",
+    "permission-group",
+    "permission-individual",
+    "photos",
+    "placeholder",
+    "play",
+    "print",
+    "products-and-services",
+    "profile",
+    "project-management",
+    "publish",
+    "quantities",
+    "redo",
+    "rfi",
+    "save",
+    "search",
+    "settings",
+    "share",
+    "sync",
+    "tag",
+    "trash",
+    "undo",
+    "unlock",
+    "upload",
+    "visible",
+    "x-close-gray"
+  ],
+  [iconSizes.PX_16]: [
+    "add",
+    "archive",
+    "assets",
+    "attachment",
+    "back",
+    "bookmark",
+    "building-ops",
+    "calendar",
+    "checkbox-checked",
+    "checkbox-indeterminate",
+    "checklist",
+    "checkmark",
+    "clock",
+    "close",
+    "close-small",
+    "collaboration",
+    "collapse-panel",
+    "comment",
+    "compare",
+    "complete",
+    "copy",
+    "cost-control",
+    "document-management",
+    "download",
+    "dropdown-caret-down",
+    "dropdown-caret-left",
+    "dropdown-caret-right",
+    "dropdown-caret-up",
+    "edit",
+    "error",
+    "expand-panel",
+    "export",
+    "external",
+    "external-link",
+    "favorite",
+    "field",
+    "file-assembly",
+    "file-document",
+    "file-generic",
+    "file-image",
+    "file-part",
+    "file-pdf",
+    "file-presentation",
+    "file-spreadsheet",
+    "file-video",
+    "file-zip",
+    "filter",
+    "filter-tokens",
+    "folder",
+    "folder-open",
+    "forward",
+    "globe",
+    "grid",
+    "help",
+    "hidden",
+    "home",
+    "info",
+    "insight",
+    "issue",
+    "item",
+    "layout",
+    "list",
+    "location",
+    "lock",
+    "model-coordination",
+    "more-vertical",
+    "paste",
+    "permission-group",
+    "permission-individual",
+    "photos",
+    "placeholder",
+    "play",
+    "print",
+    "products-and-services",
+    "profile",
+    "project-management",
+    "publish",
+    "quantities",
+    "redo",
+    "rfi",
+    "save",
+    "search",
+    "settings",
+    "share",
+    "sync",
+    "tag",
+    "trash",
+    "undo",
+    "unlock",
+    "upload",
+    "visible"
+  ]
+};
+
+export default iconKeys;
