+++ packages/icon/src/__stories__/stories.js
@@ -1,0 +1,16 @@
+import iconSizes from "../sizes";
+
+export default [
+  {
+    description: "size 24",
+    getProps: () => ({
+      size: iconSizes.PX_24
+    })
+  },
+  {
+    description: "size 16",
+    getProps: () => ({
+      size: iconSizes.PX_16
+    })
+  }
+];
